[CmdPal] Address relevance harness test review feedback
Split the pure MainListRanker unit tests out of RelevanceHarnessTests into a
dedicated MainListRankerTests class, and fold the near-duplicate ClassifyTier_*
methods into a single data-driven test. Rename the end-to-end cases from the
opaque Golden_ prefix to EndToEnd_, and build the harness ListItemMock on the
real ListItem toolkit type so it no longer carries a wall of NotImplementedException
members.

Copilot-Session: 49185697-186e-406f-b081-8c985c134274
